refactor: improve code a lot
- no more custom logger
- no more env var custom code
- better module management
