+++ Cargo.toml
@@ -22,3 +22,6 @@ regex = "1"
 structopt = "0.3"
 open = "1.4"
 git2 = "0.13"
+log = "0.4"
+flexi_logger = "0.15"
+anyhow = "1.0"
