--- src/logger.rs
@@ -1,19 +1,0 @@
-pub struct Logger {
-    verbose: bool,
-}
-
-impl Logger {
-    pub fn new(verbose: bool) -> Logger {
-        Logger { verbose }
-    }
-
-    pub fn verbose_print(&self, text: &str) {
-        if self.verbose {
-            println!("gitweb: {}", text);
-        }
-    }
-
-    pub fn print(&self, text: &str) {
-        println!("{}", text);
-    }
-}
